Telephony: Update the state of FDN Service correctly
FDN Service state was derived based on the state of PIN2. Update the
state of FDN service based on the FACILTY_LOCK messages instead.

Telephony: Change the file path to FDN,MSISDN for USIM and CSIM
Correcting the file path to EF_FDN and EF_MSISDN from Telecom DF to ADF
for USIM and CSIM.

Telephony: Correcting the file path to EF_ADN and EF_MSISDN
This fix corrects the file path to EF_ADN and EF_MSISDN.

Change-Id: I56b20d88fa02fdc9a5d1adb09a6096836a32819d
CRs-Fixed: 251497, 253195 251714.