Telephony: Try setup data call when the conditions are met in applyNewState
The data connection failed is not broadcast for the enable APN type
requests when the data is disabled from the settings menu.
Changes to try initiate the data call if the conditions are met whenever
updating the enabled/met flags of the ApnContext to make sure that the
proper data connection state is broadcast.

Change-Id: Ifba76306cfc7e0a808da44339efa01be829de65a
CRs-Fixed: 371400