Telephony(DSDS): Notify activation status when ever client is registered
Notify activation status when ever client is registered for subscription
activation status so that client will query for subscription data and
will get the slot id and application id. So subscription data will be
available with phone objects even if phone objects are changing
in case of global mode.

Change-Id: Ic535feafe790be37c013d6ab40e7f086fe045620
CRs-Fixed: 449078